Reorganised edwardMoore files
